refactor(cli): split init into api/init leaf shape
Reorganize the flat init API into the fractal leaf shape: init.mjs is now a
dispatcher + barrel that routes to api/init/run/run.mjs (the default /
--features / --all install path) and api/init/remove/remove.mjs (the
--remove-agents path), with the shared plain-logger contract in
api/init/_adapter.mjs.

Pure reorg — no behavior change. getNextSteps, noopInitLogger, and the
InitOptions / InitLogger / InitRunData types stay re-exported from the barrel,
so api/index.mjs, the CLI command handler, and the programmatic API surface are
unchanged. Human and --json output are byte-identical across init, --features
agents, --all, --remove-agents, and --json.
